feat(database): add saving and reading list data
docs: update FILE.md and SPECS.md
perf(database): decrease hash table and list max size
fix(btree): uninitialized size of get_kvs_from_btree
fix(commands): double-freeing ListNode in RPOP command

+++ headers/hashtable.h
@@ -12,9 +12,9 @@ struct FVPair {
 };
 
 struct HashTableSize {
-  uint64_t allocated; // total allocated size
-  uint64_t filled; // filled allocated block count
-  uint64_t all; // contains next values
+  uint32_t allocated; // total allocated size
+  uint32_t filled; // filled allocated block count
+  uint32_t all; // contains next values
 };
 
 struct HashTable {
